Removed all old code into old_code folder
All the old code for 2D non-tensor-product structure based filters and
transforms are relocated into old_code folder. The folder could be
removed from search path in the future. However, not all other code are
cleaned currently. Some code (such as GSM) are still calling old
functions for transforms. This should be cleaned in the future.