Changed the code structure for DT
Moved the Split Filters and Normalization code to Filters and Transforms
folders respectively. Now, the time domain DT based code is better
organized. If you just want to use the transform code and combine with
some other external algorithms, you can just addpath the folders:
1. Utilities
2. Filters
3. Transforms
This would be enough for time domain DT based transforms. The interface to
external code would be:
1. transform and inverse transform
2. CalFilterNormDT2D for calculating filter norms.
3. normcoef_dt, unnormcoef_dt for normalization.

For the frequency-based transforms, the current problem is the calculation
of filter norm is quite slow. So we still need to precalculate it, and
save in a mat file. We might(?) overcome this difficulty by calculating the
1-D filter first, and take a tensor product, in the future work.